backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
sqlmodel==0.0.22
sqlalchemy==2.0.25
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
python-multipart==0.0.9
